requirements.txt
absl-py==1.0.0
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chevron==0.14.0
cryptography==35.0.0
h11==0.12.0
idna==3.3
mustache==0.1.4
outcome==1.1.0
Pillow==8.4.0
pycparser==2.21
pycurl==7.44.1
pyjson==1.3.0
pyOpenSSL==21.0.0
PySocks==1.7.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
